requirements.txt
# Core dependencies
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8

# AI/ML
google-generativeai==0.8.5
numpy==2.0.2

# Database
psycopg2-binary==2.9.10

# Utilities
python-dotenv==1.1.0

# For PDF support
pypdf==3.17.4

# For DOCX support
docx2txt==0.8

